chore: rename walker.ts => visit.ts
